requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Babel==4.0.0
Werkzeug==3.0.3
pytest==8.2.2
pytest-flask==1.3.0
